// apps/web/src/contracts/generated/CrsSystem.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { GeographicDefinition } from "./GeographicDefinition";
import type { LocalDefinition } from "./LocalDefinition";
import type { TmDefinition } from "./TmDefinition";

/**
 * The kind of system a definition is.
 */
export type CrsSystem = { "kind": "tm" } & TmDefinition | { "kind": "geographic" } & GeographicDefinition | { "kind": "local" } & LocalDefinition;